+++ miden-lib/src/accounts/faucets/mod.rs
@@ -2,8 +2,8 @@ use alloc::string::ToString;
 
 use miden_objects::{
     accounts::{
-        Account, AccountCode, AccountId, AccountStorage, AccountStorageMode, AccountType,
-        StorageSlot,
+        Account, AccountCode, AccountId, AccountProcedureInfo, AccountStorage, AccountStorageMode,
+        AccountType, StorageSlot,
     },
     assets::TokenSymbol,
     AccountError, Felt, Word, ZERO,
@@ -55,6 +55,15 @@ pub fn create_basic_fungible_faucet(
 
     let assembler = TransactionKernel::assembler();
     let account_code = AccountCode::compile(source_code, assembler, true)?;
+
+    // TODO: Remove this manual modification once we have the ability to set sizes using the
+    // assembler.
+    let procedures = account_code
+        .procedures()
+        .iter()
+        .map(|proc| AccountProcedureInfo::new(*proc.mast_root(), proc.storage_offset(), 2).unwrap())
+        .collect();
+    let account_code = AccountCode::from_parts(account_code.mast(), procedures);
 
     // First check that the metadata is valid.
     if decimals > MAX_DECIMALS {