+++ miden-tx/tests/integration/main.rs
@@ -1,3 +1,5 @@
+extern crate alloc;
+
 mod scripts;
 mod wallet;
 
@@ -30,7 +32,7 @@ pub fn prove_and_verify_transaction(
 
     let proof_options = ProvingOptions::default();
     let prover = LocalTransactionProver::new(proof_options);
-    let proven_transaction = prover.prove(executed_transaction).unwrap();
+    let proven_transaction = prover.prove(executed_transaction.into()).unwrap();
 
     assert_eq!(proven_transaction.id(), executed_transaction_id);