+++ miden-tx/tests/integration/wallet/mod.rs
@@ -1,3 +1,5 @@
+use alloc::sync::Arc;
+
 use miden_lib::{accounts::wallets::create_basic_wallet, AuthScheme};
 use miden_objects::{accounts::AccountId, crypto::dsa::rpo_falcon512::SecretKey, Word};
 use rand_chacha::{rand_core::SeedableRng, ChaCha20Rng};