+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.6.0 (TBD)
 
+- Added `RemoteTransactionProver` struct to `miden-tx-prover` (#921).
+- [BREAKING] Changed `TransactionProver` trait to be `maybe_async_trait` based on the `async` feature (#913).
+- [BREAKING] Changed `TransactionExecutor` and `TransactionHost` to use trait objects (#897).
 - Implemented kernel procedure to set transaction expiration block delta (#897).
 - Created a proving service that receives `TransactionWitness` and returns the proof using gRPC (#881).
 - Made note scripts public (#880).
@@ -18,10 +21,14 @@
 - Added `AccountStorageHeader` (#876).
 - Implemented generation of transaction kernel procedure hashes in build.rs (#887).
 - [BREAKING] `send_asset` procedure was removed from the basic wallet (#829).
+- Implemented ability to invoke procedures against the foreign account (#882, #890, #896).
 - [BREAKING] Updated limits, introduced additional limits (#889).
 - Introduced `AccountDelta` maximum size limit of 32 KiB (#889).
 - [BREAKING] Moved `MAX_NUM_FOREIGN_ACCOUNTS` into `miden-objects` (#904).
 - Implemented `storage_size`, updated storage bounds (#886).
+- [BREAKING] Auto-generate `KERNEL_ERRORS` list from the transaction kernel's MASM files and rework error constant names (#906).
+- Implement `Serializable` for `FungibleAsset` (#907).
+- [BREAKING] Changed type of `EMPTY_STORAGE_MAP_ROOT` constant to `RpoDigst`, which references constant from `miden-crypto` (#916).
 
 ## 0.5.1 (2024-08-28) - `miden-objects` crate only